+++ src/test/java/org/perfcake/parsing/ScenarioParserTest.java
@@ -137,10 +137,10 @@ public class ScenarioParserTest {
          // Message 1 validatorIds
          final List<MessageValidator> validatorsList1 = validationManager.getValidators(mts1.getValidatorIds());
          Assert.assertEquals(validatorsList1.size(), 2, "message1 validatorIdList size");
-         Assert.assertTrue(validatorsList1.get(0).isValid(new Message("Hello, this is Stupid validator")));
-         Assert.assertFalse(validatorsList1.get(0).isValid(new Message("Hello, this is Smart validator")));
-         Assert.assertTrue(validatorsList1.get(1).isValid(new Message("Hello, this is happy validator :)")));
-         Assert.assertFalse(validatorsList1.get(1).isValid(new Message("Hello, this is sad validator :(")));
+         Assert.assertTrue(validatorsList1.get(0).isValid(null, new Message("Hello, this is Stupid validator")));
+         Assert.assertFalse(validatorsList1.get(0).isValid(null, new Message("Hello, this is Smart validator")));
+         Assert.assertTrue(validatorsList1.get(1).isValid(null, new Message("Hello, this is happy validator :)")));
+         Assert.assertFalse(validatorsList1.get(1).isValid(null, new Message("Hello, this is sad validator :(")));
 
          // Message 2
          final MessageTemplate mts2 = messageStore.get(1);
@@ -157,8 +157,8 @@ public class ScenarioParserTest {
          // Message 2 validatorIds
          final List<MessageValidator> validatorsList2 = validationManager.getValidators(mts2.getValidatorIds());
          Assert.assertEquals(validatorsList2.size(), 1, "message2 validatorIdList size");
-         Assert.assertTrue(validatorsList2.get(0).isValid(new Message("Go for fishing!")));
-         Assert.assertFalse(validatorsList2.get(0).isValid(new Message("Go for mushroom picking! There are no Fish.")));
+         Assert.assertTrue(validatorsList2.get(0).isValid(null, new Message("Go for fishing!")));
+         Assert.assertFalse(validatorsList2.get(0).isValid(null, new Message("Go for mushroom picking! There are no Fish.")));
 
          // Message 3
          final MessageTemplate mts3 = messageStore.get(2);